+++ packages/speculate/src/libs/speculate-links.ts
@@ -1,4 +1,5 @@
 import type { SpeculationTriggers, SpeculationActions } from "../types.js";
+import validConnection from "../utils/valid-connection.js";
 
 const C: {
 	moderateEvents: Array<keyof HTMLElementEventMap>;
@@ -186,33 +187,6 @@ const shouldPreload = (props: {
 	} catch (_) {
 		return false;
 	}
-};
-
-/**
- * Connection check
- */
-const validConnection = (): boolean => {
-	if (!navigator.onLine) {
-		console.warn("The device is offline, speculate library not initialised.");
-		return false;
-	}
-	if ("connection" in navigator) {
-		const connection = navigator.connection;
-		// @ts-expect-error
-		if (connection?.saveData) {
-			console.warn("Save-Data is enabled, speculate library not initialised.");
-			return false;
-		}
-		// @ts-expect-error
-		if (/(2|3)g/.test(connection?.effectiveType)) {
-			console.warn(
-				"2G or 3G connection is detected, speculate library not initialised.",
-			);
-			return false;
-		}
-	}
-
-	return true;
 };
 
 /**